# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
...
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: getSession
      tags: [Sessions]
      responses:
        "200":
          description: Session details.
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteSession
      tags: [Sessions]
      responses:
        "202":
          description: Session sandbox deletion was requested; durable history is retained.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSessionResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/workspace/files:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: listWorkspaceFiles
      tags: [Workspace]
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: ""
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A page of entries in a workspace directory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceEntryPage"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/workspace/file:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: readWorkspaceFile
      tags: [Workspace]
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: UTF-8 workspace file contents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceFile"
        default:
          $ref: "#/components/responses/Error"
    put:
      operationId: writeWorkspaceFile
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceFileWriteInput"
      responses:
        "200":
          description: The updated UTF-8 workspace file.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceFile"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/workspace/file/diff:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: readWorkspaceDiffFile
      tags: [Workspace]
      description: Read one cloud workspace file with its bounded unified diff.
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Docker workspace file review details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDiffFileDetail"
        "501":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/workspace/diff:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: getWorkspaceDiff
      tags: [Workspace]
      description: Changed-file summary and line counts relative to the session compare base.
      responses:
        "200":
          description: Workspace changes relative to the session compare base.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDiff"
        "501":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport, and
        `worker:notification` for durable agent notification intake.
  parameters:
...
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    AgentCapability:
      type: string
      enum:
        - interface.chat
        - interface.tui
        - model.catalog
        - model.custom
        - attachments
        - browser.preview
        - review.execute
        - session.resume
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, code, message, requestId]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    PageInfo:
      type: object
      additionalProperties: false
      required: [hasMore]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: string
...
    SessionKind:
      type: string
      enum: [worker, orchestrator]
    SessionMode:
      type: string
      enum: [read-only, standard, trusted]
    SessionActivityState:
      type: string
      enum: [active, idle, waiting_input, blocked, exited]
    SessionStatus:
      type: string
      enum:
        - working
        - needs_input
        - pr_open
        - draft
        - review_pending
        - ci_failed
        - changes_requested
        - approved
        - mergeable
        - merged
        - exited
        - idle
        - terminated
        - no_signal
    Turn:
      type: object
      additionalProperties: false
      required: [id, sessionId, userMessageSequence, state, attemptCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        userMessageSequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        state:
          type: string
          enum: [queued, provisioning, running, cancel_requested, completed, failed]
        attemptCount:
          type: integer
          minimum: 0
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DeleteSessionResponse:
      type: object
      additionalProperties: false
      required: [session]
      properties:
        session:
          type: object
          additionalProperties: false
          required: [id, desiredState]
          properties:
            id:
              type: string
              format: uuid
            desiredState:
              type: string
              const: deleted
    SessionInterfaceMode:
      type: string
      enum: [tui, chat]
...
    Session:
      type: object
      additionalProperties: false
      required:
        - id
        - orgId
        - projectId
        - kind
        - harness
        - displayName
        - branch
        - mode
        - deniedCommands
        - interfaceMode
        - activityState
        - status
        - runtimeConnected
        - isTerminated
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SessionKind"
        harness:
          type: string
        displayName:
          type: string
        branch:
          type: string
        mode:
          $ref: "#/components/schemas/SessionMode"
        model:
          type: string
          description: >-
            Coding-agent model the session launched with; empty means the harness
            default.
        deniedCommands:
          type: array
          items:
            type: string
        interfaceMode:
          $ref: "#/components/schemas/SessionInterfaceMode"
        activityState:
          $ref: "#/components/schemas/SessionActivityState"
        status:
          $ref: "#/components/schemas/SessionStatus"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/AgentCapability"
        runtimeConnected:
          type: boolean
        runtimeState:
          type: string
        runtimeError:
          type: string
        activeTurn:
          $ref: "#/components/schemas/Turn"
        isTerminated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    WorkspaceEntry:
      type: object
      additionalProperties: false
      required: [name, path, isDir, size, mode, modTime]
      properties:
        name:
          type: string
        path:
          type: string
        isDir:
          type: boolean
        size:
          type: integer
          format: int64
        mode:
          type: string
        modTime:
          type: string
          format: date-time
    WorkspaceEntryPage:
      type: object
      additionalProperties: false
      required: [path, items, page]
      properties:
        path:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceEntry"
        page:
          $ref: "#/components/schemas/PageInfo"
    WorkspaceFile:
      type: object
      additionalProperties: false
      required: [path, content, size]
      properties:
        path:
          type: string
        content:
          type: string
        size:
          type: integer
          format: int64
    WorkspaceFileWriteInput:
      type: object
      additionalProperties: false
      required: [path, content]
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 4096
        content:
          type: string
          maxLength: 1048576
    WorkspaceFileStatus:
      type: string
      enum: [unmodified, modified, added, deleted, renamed, untracked, copied, changed]
    WorkspaceDiffFile:
      type: object
      additionalProperties: false
      required: [path, status, additions, deletions, binary]
      properties:
        path:
          type: string
        oldPath:
          type: string
        status:
          $ref: "#/components/schemas/WorkspaceFileStatus"
        staged:
          type: string
        unstaged:
          type: string
        additions:
          type: integer
          minimum: 0
        deletions:
          type: integer
          minimum: 0
        binary:
          type: boolean
    WorkspaceDiffFileDetail:
      type: object
      additionalProperties: false
      required: [path, status, additions, deletions, size, binary, deleted, content, contentTruncated, diff, diffTruncated]
      properties:
        path:
          type: string
        status:
          $ref: "#/components/schemas/WorkspaceFileStatus"
        additions:
          type: integer
          minimum: 0
        deletions:
          type: integer
          minimum: 0
        size:
          type: integer
          format: int64
          minimum: 0
        binary:
          type: boolean
        deleted:
          type: boolean
        content:
          type: string
        contentTruncated:
          type: boolean
        diff:
          type: string
        diffTruncated:
          type: boolean
    WorkspaceDiff:
      type: object
      additionalProperties: false
      required: [status, unstaged, staged, combined, diffBaseRef, files, untrackedFiles, truncated]
      properties:
        status:
          type: string
        unstaged:
          type: string
        staged:
          type: string
        combined:
          type: string
        diffBaseRef:
          type: string
        diffBaseSha:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceDiffFile"
        untrackedFiles:
          type: array
          items:
            type: string
        truncated:
          type: object
          additionalProperties: false
          required: [combined, stats]
          properties:
            combined:
              type: boolean
            stats:
              type: boolean
...